+++ inc/help.sh
@@ -84,7 +84,7 @@ print_help() {
 		explore)
 			echo "Command usage: $command [OPTIONS] [PATH]"
 			echo
-			echo "Open file browser to explore backuped files"
+			echo "Open file browser to explore backed up files"
 			echo "If no path is specified, it will open the root backup folder."
 			echo "Warning: This feature does not detect old renamed/moved files yet."
 			echo